+++ client/src/pages/admin/scenarios/[scenarioId]/edit.tsx
@@ -48,6 +48,7 @@ const UpdateScenarioPage: React.FC = () => {
           onSuccess: () => {
             toast.success('Your changes were successfully saved.');
             queryClient.invalidateQueries(['scenariosList']);
+            queryClient.invalidateQueries(['scenario', data.id]);
           },
           onError: (error: ErrorResponse) => {
             const { errors } = error.response?.data;
@@ -176,13 +177,6 @@ const UpdateScenarioPage: React.FC = () => {
                             <DotsVerticalIcon className="w-4 h-4" />
                           </Dropdown.Button>
                           <Dropdown.Items>
-                            {/* <Dropdown.Item>
-                              <Link
-                                href={`/admin/scenarios/${data.id}/interventions/${intervention.id}/edit`}
-                              >
-                                <a className="block p-2 text-sm">Edit</a>
-                              </Link>
-                            </Dropdown.Item> */}
                             <Dropdown.Item>
                               <Link
                                 href={`/admin/scenarios/${data.id}/interventions/${intervention.id}/edit`}